In-the-wild crawl fixes
We've observed that an SSL certificate is sometimes thrown by the
webhost itself and not our local ca certificate. In other words, the
ca certificate can be correctly installed in the system keychain and yet
we'll still fail with the SSL error. Modify this error to be a warning
in the cases where users still want to continue.

Also cleanup the CLI manually since some resources are leaking and the
process won't end on its own.